[Fix/Delete] Reorganização de arquivos e métodos
