--- backend/src/main/java/com/produtos/apirest/repository/ProfitPerDayRepository.java
@@ -1,13 +1,0 @@
-package com.produtos.apirest.repository;
-
-import com.produtos.apirest.models.ProfitPerDay;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProfitPerDayRepository extends JpaRepository<ProfitPerDay, Long> {
-
-    ProfitPerDay findById(long id);
-
-
-    Boolean existsById(long id);
-
-}
